requirements.txt
snowflake-connector-python
elasticsearch
requests